fix symlinks

+++ scripts/symlinks.sh
@@ -3,12 +3,29 @@
 dotfiles="dotfiles"
 
 function linkDotfile {
-  echo "Creating new symlink: ${2}"
-  ln -sfn ${1} ${2}
+
+  if [ ! -L ${2} ]; then
+    if [-f "${2}" ]; then
+      # Existing file
+      echo "Backing up existing file: ${2}"
+      mv ${2}{,.${dateStr}}
+
+    elif [  -d "${2}" ]; then
+      # Existing dir
+      echo "Backing up existing dir: ${2}"
+      mv ${2}{,.${dateStr}}
+    fi
+  fi
+
+  if [ ! -e ${2} ]; then
+    echo "Generate symlink for ${2}}"
+    ln -sfn ${1} ${2}
+  fi
+  
 }
 
 linkDotfile ~/$dotfiles/fonts ~/.fonts
-linkDotfile ~/$dotfiles/config/zsh/.zshrc ~/.zshrc
+linkDotfile ~/$dotfiles/config/zsh/zshrc ~/.zshrc
 linkDotfile ~/$dotfiles/config/compton.conf ~/.config/compton.conf
 linkDotfile ~/$dotfiles/config/dunst ~/.config/dunst
 linkDotfile ~/$dotfiles/config/i3 ~/.config/i3
